requirements.txt
python-dotenv==1.0.0
requests==2.30.0
python-telegram-bot==13.15
google-cloud-dialogflow==2.23.2
google-cloud==0.34.0
google-cloud-api-keys==0.5.3